Keep Waiting Discord probe off the guild-role graph.
Extract the official-guild membership GET so waitingSummary does not
pull role-sync into the platform startup bundle, stop exporting the
coarse checklist skip list, and raise the reviewed platform budget.
